backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment configuration
python-dotenv==1.0.0

# Database - MongoDB
motor==3.3.2
pymongo==4.6.0

# Cache - Redis
redis[hiredis]==5.0.1

# AWS Services
boto3==1.29.7
botocore==1.32.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1  # For TestClient

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging
python-json-logger==2.0.7

# External APIs
openai==1.3.7  # For Whisper and GPT-4 (Epic 2 and 3)

# Celery for async task processing
celery[sqs]==5.3.4
kombu==5.3.4

# Fuzzy string matching for entity resolution (Story 3.3)
rapidfuzz==3.5.2

# Vector search - OpenSearch (Epic 4)
opensearch-py==2.4.2
requests-aws4auth==1.2.3

# Authentication - JWT validation (Story 5.1)
python-jose[cryptography]==3.3.0
pyjwt[crypto]==2.8.0
cryptography==41.0.7
# httpx==0.25.1 already listed in testing section

# System metrics (Story 5.6)
psutil==5.9.6
